Make twophase tolerance test resilient.
CI can get into the weird state due to which recovery can take a long time,
affecting mirror promotion time. If the mirror takes a long time for promotion,
master can panic if can't complete commit-prepared in time. Hence, make the test
more resilient by bumping the number of retries to 10.

Also, to make sure gprecoverseg doesn't fail due to "can't start transaction in
BEGIN", add gucs to make sure the retries are higher if the command fails if the
segment is recovering.