Relax filerep resync logic to send change tracking entries
Issue is that filerep assumed that the LSN of page is always greater than LSN
of the change tracking log for that page. This assumption was broken (in the
specific case of hint bits) in 5.X by heap checksums back port from upstream.
The breach in the assumption led to filerep resync incorrectly not syncing
certain blocks from primary to mirror.

As part of the heap-checksum back port, we introduced XLOG_HINT records to
capture full page images in XLOG. A transaction that sets hints bits on a page
emits XLOG_HINT record containing full image of that page. This is to avoid
false alarms during checksum validation, in the event of a torn page write.
After emitting XLOG_HINT record, the code ({{MarkBufferDirtyHint()}}) sets the
page's LSN only if the page is not already marked dirty. The implication of
this logic is that a page's LSN may remain lower than the LSN of the most
recent XLOG record emitted for that page (which is also the LSN recorded in
filerep change tracking log).

Co-authored-by: David Kimura <dkimura@pivotal.io>