Set motion node's rescannable field to false.
This commit fix the github issue 5628.

We can not create a motion directly in the RHS of Nested loop
join, because Motion node is not rescannable.

In the previous codes, we set the rescannable field the value
same as subpath's even if the RHS is SegmentGeneral or SingleQE.
We fix this by setting the motion's rescannable to false directly.

Co-authored-by: Shujie Zhang <shzhang@pivotal.io>
Co-authored-by: Zhenghua Lyu <zlv@pivotal.io>