Return NULL pointer rather than implied NULL via false
CdbDispatchDtxProtocolCommand() is defined to return struct pg_result **, but
returned false in an error path which forces zero imply a NULL pointer. The
caller of this function has another error check and wouldn't actually read
the returned value, but it should still be doing the right thing. Fix by
returning NULL instead.  This resolves the followig clang compiler warning:

  cdbdisp_dtx.c:157:10: warning: expression which evaluates to zero treated as
                        a null pointer constant of type 'struct pg_result **'
                        [-Wnon-literal-null-conversion]
                  return false;
                         ^~~~~
  ../../../../src/include/c.h:195:15: note: expanded from macro 'false'
  #define false   ((bool) 0)
                  ^~~~~~~~~~