Fix assertion failure trying to build a Motion on a non-hashable column.
Be more careful to not build a Redistribute Motion on an expression that's
not GPDB-hashable.

This is a backport of commit a031b9c014 from master. It did not apply
cleanly to 5X_STABLE, which is why this leaves one of the test queries
still failing, but now with a clean elog() rather than an assertion
failure. The error message is memorized in the output, and opened a
separate github issue to track it, but I'd like to get at least this much
pushed now.