ci: fix race condition in gptransfer ccp multi-cluster tests
A race condition was occuring because the set of concourse tasks that
create the second of two CCP clusters was expecting to find and use the
`terraform` volume. The `terraform` volume is expected to only be created
and used in the first set of tasks for the first CCP cluster. If the
first set of tasks did not complete before the second set then there was
the potential for the `terraform` volume to not exist yet. This causes
the job to error in concourse.

The fix is to correct the mistake of the second set of tasks using the
wrong volume. They should only be using the `terraform2` volume. This
completely removes the potential for the race condition.

Authored-by: Kris Macoskey <kmacoskey@pivotal.io>