Explicitly select the correct side in restrictinfo for copathkey generation in cdbpath_match_preds_to_partkey_tail().
Just like what we do in cdbpath_partkeys_from_preds(). This is more
friendly for code reading and also is better for code maintenance.
Previous code could potentially introduce bugs since it's relids, not
equivalenceclass, that could determine the side for comparison accurately.

I could not create a test case for this unfortunately.