Support equijoin on relabeled distribution keys
Distribution key joins on pathkeys where the parser invokes an implicit
cast via a RelabelType node would introduce a redistribution motion due
to the join not being recognized as a colocated equijoin.

As an example, consider the following relation

    create table test (a varchar(10), b varchar(10)) distributed by (a);

The below query will introduce an implicit cast on a.a and b.a to TEXT by
the parser. This means that the pathkey EquivalenceClass is on VARCHAR but
the RestrictInfo ECs are for TEXT.

    select * from test a, test b where a.a=b.a;

Fix by following the EC members to their basetypes to allow for colocated
equijoins in these cases.

Affected testfiles updated, and one affected testcase slightly tweaked to
ignore less and generate less uninteresting output in the .out file.