Remove remaining unused pieces of wal_consistency_checking.
Most of the backup block related modification for providing the
wal_consistency_checking was removed as part of 9.3 merge. This was mainly done
to avoid merge conflicts. The masking functions are still used by
gp_replica_check tool to perform checking between primary and mirrors. But the
online version of checking during each replay of record was let go. So, in this
commit cleaning up remaining pieces which are not used. We will get back this in
properly working condition when we catch up to upstream.