Rename variable to avoid risk of type collision
The comment states that "small" might be defined by socket.h, and while
thats not true for all versions of sys/socket.h, it's still not a good
name to use as it's common in Windows headers (should we ever revive a
Windows port). Renaming to a non-colliding name is a small price to pay
to avoid subtle bugs, so rename and remove the preprocessor dance.