Remove bkuprestore test
The bkuprestore test was imported with the source code, but has never
been used and hasn't worked in a long time. Rather than trying to save
this broken mess, let's remove it and start fresh with a pg_dump TAP
test which is a much better way to test backup/restore.