Use APR from apt
The Apache Portable Runtime (APR) is not a direct dependency of the
Greenplum server, but it's depended upon by `gpfdist` and `gpperfmon`.
Judging from the code, we don't seem to depend on any (very) particular
version of APR, as long as it's something recent-ish (say 1.5.0+).

Commit 168f612a73e (#58) started our Travis CI config with directly
downloading APR from Apache releases. This was probably due to the fact
that Travis in 2015 was still running Ubuntu 12.04 Precise that shipped
too old of a version of APR. Over the years, updating APR has been a
recurrent semiannual ritual, see commits 530db5ea71e (#5873),
27adcf92a13, 19251f3ff72, 5a9885f4f84, 2c68384e510.

This patch ends that fun madness by installing `libapr1-dev` from apt.
It turns out -- interestingly enough -- APU (a sister library of APR)
isn't needed in Travis: when we started using Travis (in commit
168f612a73e #58), we were building APU, but APU is really _only_ a
dependency of `gpperfmon`, which we _never_ built in Travis.