Generate name of the index backing a constraint
ALTER TABLE ... ADD CONSTRAINT generates the name of the index to be
created.  This patch uses the same logic to generate the name during
ugprade, so that OIDs can be pre-assigned to the right name.
Exclusion constraints do not have a SQL syntax that allows the index
name to be specified, so we do so via a udf.  This compilcates the
fix, because the udf must be executed in the upgraded instance.  So we
cannot execute the udf in the old instance and get the constraint
index name.  While we could have done without the udf for primary and
unique constraints with "ALTER TABLE ... ADD CONSTRAINT USING INDEX"
syntax, we decided to use the udf for all three to make this procedure
uniform.

It is no longer necessary to drop public.newpart table from regression
database before upgrade.

Co-authored-by: David Krieger <dkrieger@pivotal.io>