Improve Oid sync logic for both QD and QE in wraparound cases
When Oid count between the QD and QE were not synced as expected, there are
cases where object creation could take a long time due to Oid synchronizing.

These are the two that we found:
1. If the QD wrapped around and the QE did not, the QD syncing with the highest
QE Oid counter value could result in a long loop. Additionally, there's a hack
in how the QD advances its Oid counter above the highest QE Oid counter value
where it will increment the Oid counter 10 more times to be *safe*. However,
this hack introduces holes in the Oid count and makes this bug more probable.
2. The QE advances its Oid counter when creating objects with dispatched
pre-assigned Oids. If the pre-assigned Oid dispatched by the QD was not from
wraparound and the QE had already wrapped around, it would result in a long
loop for the QE to synchronize with the QD.

To prevent long Oid synchronization loops, we add logic to handle wraparound
cases in both QD and QE.

This can be easily reproduced by running `pg_resetxlog -o <Oid>` to set the Oid
counter value on QD and/or QE to simulate the scenarios.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>
Co-authored-by: Jesse Zhang <sbjesse@gmail.com>