Make fts_error test deterministic.
The fts_error tests fails sometimes, as `gp_wait_until_triggered_fault()`
functions behavior is to return after the fault location is hit. It can't
communicate whether the fault action was taken or not. Due to this, the test
sometimes used to fail if checking for locks happened before FTS process
completed error handling.

So, adding fts scan to be performed which should make sure the new FTS process
is spanned and then only check if any locks remain.

Co-authored-by: Joao Pereira <jdealmeidapereira@pivotal.io>