Only perform SIRV function transformation in the target list of DML
The SIRV transformation was preventing some PL/pgSQL expressions from
being treated as "simple". That's a lost optimization, and it also caused
some extra context lines being printed in error messages, compared to
upstream. While those context lines are harmless, and I wouldn't expect
GPDB necessarily to totally match upstream behavior on when and how error
context is printed, it's always nicer if we can match upstream, for the
sake of avoiding differences in expected output of regression tests.

There's one interesting test query in 'qp_resource_queue', where this
causes functions in target list to be called more times than before.
The case involves test functions, as well as a set-returning-function
(generate_series), being used in the same target list. The new behavior,
which is what PostgreSQL also does, is to call all the functions for
each row returned by generate_series(), whereas before this commit, they
were only called once. The behavior with set-returning-functions in
target list is pretty whacky in many ways, so I don't think there's any
principled way to say which is more correct, but following the upstream
always seems reasonable. However - and this is perhaps not so reasonable -
the behavior is now different in a DML statement and a plain SELECT, as
we still do the SIRV transformation in DML statements, but I think we
can live with that.