Modify the default value of Debug_dtm_action_segment.
This GUC had default value as 1. If any tests this guc fails to set it (means
even if the SET command fails in test to change the value) the default kicks in
and segment with content id 1 is affected. Best to have this guc point to
invalid segment. Any tests using the same must explicitly set the same to
intended value.

Ideally, this guc based approach must go away and fault injector framework must
be used for the same. But that's for other commit.
Adjust distributed_transactions to set debug_dtm_action_segment=1.