Make udf_exception_blocks_panic_scenarios deterministic.
The test was setting `debug_dtm_action_segment=1` means take action of segment
with content id 1. But when checking for if segment recovered fine or not
checked for content id 0 using `0U: select 1;`.

So, lets make the test deterministic by using content 0 for PANIC as well and
content 0 for checking if recovery completed as well.