Add infrastructure to start using fakes
In this commit the infrastructure to add fakes is added.
Now we can create fake implementations of our objects in
`src/test/unit/test_doubles/path/to/file/realfilename_fake.c`

When we use them we need to add the following to the makefile:

```
shmem_fake.t: $(TEST_DOUBLE_DIR)/backend/storage/ipc/shmem_fake.o
```

This will include the fake instead of the real implementation of
the specific object.

As a proof of concept a shmem_fake was added with some tests that cover
both the fake and the real implementation.