Disallow array types as hashable types
Previously if the attribute was an array type such as numeric[], float[] etc,
the OID for those attributes would be considered as ANYARRAYOID instead of
their original type. This would enable the hashDatum() to hash the data for
that particular attribute since it only knows how to hash ANYARRAYOID. But this
resulted in incorrect behavior. So for now disallow all attributes which are
array types from being hashed until the hashDatum() is fixed.

Fixes github issue #5467.

Co-authored-by: Sambitesh Dash <sdash@pivotal.io>