Fix gpnetbench function prototypes
With the gpnetbench Makefile building with the same warninglevels as
the rest of the tree, the code no longer compiled without warnings.
Fix by adding the requisite prototype, and make them static while at
it because why not. Also reformatted the function heads to match the
code style used in the tree while at it.