Allow EXPLAIN ANALYZE when auto_explain is on
Enable the user to perform EXPLAIN ANALYZE when auto_explain is on. When such query is executed, auto_explain is disabled, as auto_explain uses 'queryDesc->showstatctx' which will be initialized (and used) by EXPLAIN ANALYZE itself.

To distinguish between EXPLAIN ANALYZE and other queries, a new execution flag, 'EXEC_FLAG_ANALYZE', is added to executor/executor.h.
This flag is set by 'ExplainOnePlan()' before the start of a query. Then, it is checked by auto_explain, and if present, auto_explain is disabled (in fact, the flag is checked in every query, and the state of auto_explain is changes according to that flag).

Another solution would be to introduce another 'CdbExplain_ShowStatCtx' object in QueryDesc.