COPY now correctly reports an error if it fails on a segment.
Previously, if COPY failed on a segment, it would silently "succeed"
but in reality failed.  That behavior would have been interpreted
by a user as "COPY passed but there was no data to COPY" whereas
in reality there was an error that occurred.  This commit fixes that
and adds a test using the fault injector..

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>