Remove "Rows Out" processing in test explain parsing
The Rows out.. output was removed from EXPLAIN in the recent merge
with PostgreSQL 9.  This output remained in the expected testfiles
though, and that worked due to the explain Perl module parsing out
the offending line from the resulting diff operation. This removes
the parsing and all the now-outdated rows from the expected files.