gpaddmirrors test: Cleanup cluster at end of each test.
We have observed intermittent failures in gpaddmirrors test with the error:
`pg_basebackup: could not connect to server: FATAL:  number of requested standby
connections exceeds max_wal_senders (currently 1)`.

This implies that a there was a connection to the Wal Sender process when the
pg_basebakeup is initiated for the new mirror.

Looking at the gpaddmirror tests, we see that when creating a new cluster, the
tests don't cleanup all the segments/hosts of the previous test.  Thus leaving
mirrors around that do try to connect to the primary of the new cluster.
However, the Wal Sender of the new primary sends the following error to the old
mirror:

```... "ERROR","XX000","database system identifier differs between the primary
and standby (libpqwalreceiver.c:162)","The primary's identifier is
6613767102099215083, the standby's identifier is
6613766724142160379.",,,,,,0,,"libpqwalreceiver.c",162,"Stack trace: '''

During the small window that the Wal Sender is connected to the old mirror, a
pg_basebackup from the new mirror would get the FATAL error that we are seeing
in our test output.

More evidence that this is occuring is that during a green test, as in one that
has had pg_basebackup connnect successfully, we see the "orphaned" mirror
attempting to connect and failing.

This commit has each scenario stop its own database.

Co-authored-by: Jim Doty <jdoty@pivotal.io>
Co-authored-by: Shoaib Lari <slari@pivotal.io>