Workaround flaky test udf_exception_blocks_panic_scenarios
Test exhibits unexpected behavior which seems to cause a gang reset and
lost `debug_dtm_*` GUC settings. We noticed that FTS probe led to gang
reset whenever we fail with following diff:

```
DROP TABLE IF EXISTS employees;
 DROP
  select test_protocol_allseg(1, 2,'f');
  -ERROR:  DTX RollbackAndReleaseCurrentSubTransaction dispatch failed
  -CONTEXT:  PL/pgSQL function
  test_protocol_allseg(integer,integer,character) line 18 during
  exception cleanup
  +test_protocol_allseg
  +--------------------
  +
  +(1 row)
```

We believe this is due to interactions between FTS and frequent UDF
execeptions. Until a more solid fix is identified, a simple workaround
is to turn off FTS during this test since FTS is not required to
exercise the original intent behind this test.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>